+++ modules/Backends/lib/libfirst.cpp
@@ -99,19 +99,19 @@ double nastyExample(int count, ...)
 }
 
 // an example function taking a 1D array as an argument
-int arrayarg_1D(double arg[10])
+double arrayarg_1D(double arg[10])
 {
   return arg[0]*arg[9]; 
 }
 
 // an example function taking a 1D array as an argument
-int arrayarg_2D(double arg[10][10])
+double arrayarg_2D(double arg[10][10])
 {
   return arg[0][0]*arg[9][0]*arg[9][9]; 
 }
 
 // an example function taking a 1D array as an argument
-int arrayarg_3D(double arg[10][10][10])
+double arrayarg_3D(double arg[10][10][10])
 {
   return arg[0][0][0]*arg[9][0][0]*arg[9][9][9]; 
 }
